chore: move older pages to pages/legacy
