+++ dropbox-hacking-photo-manager-client/src/index.tsx
@@ -13,14 +13,14 @@ import { context as routingContext, type Router } from "@/hooks/useRouter";
 import * as rxRecordFeedContext from "./context/rx/rxRecordFeedContext";
 import * as additionalFeeds from "./context/rx/additionalFeeds";
 import * as thumbnailLoaderContext from "./context/thumbnails";
-import Day from "./day";
-import Calendar from "./days/calendar";
-import ListOfDays from "./days/listOfDays";
+import Day from "@/pages/legacy/day";
+import Calendar from "@/pages/legacy/days/calendar";
+import ListOfDays from "@/pages/legacy/days/listOfDays";
 import logRender from "@/logRender";
-import Photo from "./photo";
-import ClosestTo from "./closest-to/index";
-import Month from "./month";
-import Year from "./year";
+import Photo from "@/pages/legacy/photo";
+import ClosestTo from "@/pages/legacy/closest-to/index";
+import Month from "@/pages/legacy/month";
+import Year from "@/pages/legacy/year";
 import BasicCounts from "@/pages/next-gen/basic-counts";
 import NGDaysNoSamples from "@/pages/next-gen/list-of-days/without-samples";
 import NGDayFiles from "@/pages/next-gen/day/files";
