+++ dropbox-hacking-photo-manager-client/src/pages/legacy/day/index.tsx
@@ -9,10 +9,10 @@ import useVisibilityTracking from "../days/useVisibilityTracking";
 import logRender from "@/logRender";
 import EditableTextField from "./editableTextField";
 import PhotoTile from "./photoTile";
-import { useRxFeedsViaMultiplexer } from "../context/rx/rxRecordFeedContext";
+import { useRxFeedsViaMultiplexer } from "@/context/rx/rxRecordFeedContext";
 import { map } from "rxjs";
-import { useLatestValue } from "../context/rx/useLatestValue";
-import { useAdditionalFeeds } from "../context/rx/additionalFeeds";
+import { useLatestValue } from "@/context/rx/useLatestValue";
+import { useAdditionalFeeds } from "@/context/rx/additionalFeeds";
 import SamePageLink from "@/components/samePageLink";
 
 const DayWithData = ({
