+++ dropbox-hacking-photo-manager-client/src/pages/legacy/day/photoTile.tsx
@@ -1,8 +1,8 @@
 import { GPSLatLong, Photo } from "dropbox-hacking-photo-manager-shared";
 import * as React from "react";
 
-import { useThumbnailLoader } from "../context/thumbnails";
-import { useThumbnail } from "../context/thumbnails/useThumbnail";
+import { useThumbnailLoader } from "@/context/thumbnails";
+import { useThumbnail } from "@/context/thumbnails/useThumbnail";
 import logRender from "@/logRender";
 import SamePageLink from "@/components/samePageLink";
