+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/emittableSubscribable.ts
