+++ dropbox-hacking-photo-manager-client/src/pages/legacy/photo/index.tsx
@@ -3,8 +3,8 @@ import * as React from "react";
 import { useEffect, useMemo, useState } from "react";
 
 import logRender from "@/logRender";
-import { useLatestValue } from "../context/rx/useLatestValue";
-import { useAdditionalFeeds } from "../context/rx/additionalFeeds";
+import { useLatestValue } from "@/context/rx/useLatestValue";
+import { useAdditionalFeeds } from "@/context/rx/additionalFeeds";
 import SamePageLink from "@/components/samePageLink";
 import EditableTextField from "../day/editableTextField";
