+++ dropbox-hacking-photo-manager-client/src/pages/legacy/stats/index.tsx
@@ -3,11 +3,11 @@ import { useEffect, useMemo, useState } from "react";
 import { map, Observable } from "rxjs";
 
 import logRender from "@/logRender";
-import { useRxFeedsViaMultiplexer } from "../context/rx/rxRecordFeedContext";
-import type { ImageAndMaybeDelta } from "../context/rx/rxFeedClient";
-import Navigate from "../days/navigate";
-import { useLatestValue } from "../context/rx/useLatestValue";
-import { useAdditionalFeeds } from "../context/rx/additionalFeeds";
+import { useRxFeedsViaMultiplexer } from "@/context/rx/rxRecordFeedContext";
+import type { ImageAndMaybeDelta } from "@/context/rx/rxFeedClient";
+import Navigate from "@/pages/legacy/days/navigate";
+import { useLatestValue } from "@/context/rx/useLatestValue";
+import { useAdditionalFeeds } from "@/context/rx/additionalFeeds";
 
 const Stats = () => {
   const feeds = useRxFeedsViaMultiplexer();
