+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/fileRev/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 
 import logRender from "@/logRender";
-import Navigate from "@/days/navigate";
+import Navigate from "@/pages/legacy/days/navigate";
 import type { FileRevResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
 import SamePageLink from "@/components/samePageLink";
